src/broker_api/requirements.txt
databento==0.47.0
databento-dbn==0.25.0
fastapi==0.115.6
grpcio==1.68.0
grpcio-tools==1.68.0
ipykernel==6.29.5
ib_async==1.0.3
python-dotenv==1.0.1
beautifulsoup4>=4.12
ipython==8.30.0
json5==0.10.0
matplotlib==3.9.3
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.2.0
pandas==2.2.3
pandas_market_calendars==4.4.2
protobuf==5.29.0
pydantic==2.10.3
pytz==2024.2
requests==2.32.3
schwab-py
scikit-learn==1.6.0
scipy==1.14.1
seaborn==0.13.2
setuptools==75.6.0
six==1.17.0
sympy==1.13.1
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.34.0
websocket-client==1.8.0
